% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsanta.R
\name{secretsanta}
\alias{secretsanta}
\title{secretsanta}
\usage{
secretsanta(sender_email, sender_pwd)
}
\arguments{
\item{sender_email}{Email address from which the secret santa assignments will be communicated}

\item{sender_pwd}{Password to above email address}
}
\description{
Facilitate secret santa gift exchange with user input of participant names and email addresses
}
\examples{
example1
}
